requirements.txt
-i https://pypi.org/simple
anyio==3.6.1 ; python_full_version >= '3.6.2'
click==8.1.3 ; python_version >= '3.7'
dnspython==2.2.1 ; python_version >= '3.6' and python_version < '4.0'
email-validator==1.3.0
fastapi==0.85.0
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
pydantic[email]==1.10.2
pymongo==4.2.0
python-dotenv==0.21.0
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.20.4 ; python_version >= '3.7'
typing-extensions==4.4.0 ; python_version >= '3.7'
uvicorn==0.18.3
